requirement.txt
google-generativeai
SpeechRecognition
python-dotenv
fuzzywuzzy
python-Levenshtein
pyaudio
elevenlabs
pyttsx3
pygame